I have merge the google_login.py file into server.py file
