+++ api/server.py
@@ -2,6 +2,16 @@ from flask import Flask, render_template, jsonify
 import json
 from collections import OrderedDict
 
+import os
+import pathlib
+
+import requests
+from flask import Flask, session, abort, redirect, request
+from google.oauth2 import id_token
+from google_auth_oauthlib.flow import Flow
+from pip._vendor import cachecontrol
+import google.auth.transport.requests
+
 data = []
 with open('output.json', 'r') as file:
     data = json.load(file)
@@ -42,8 +52,80 @@ def searchFile(filter):
 print(searchFile('TCS-101'))
 
 app = Flask(__name__)
+app.secret_key = "ExamArchive.com"
 cources = []
 
+
+# ////////////////////////////////////////////////// GOOGLE LOGIN CODE /////////////////////////////////////////////////////////
+
+os.environ["OAUTHLIB_INSECURE_TRANSPORT"] = "1"
+
+# ------------> Google Client Id generated from Google Oauth for our end <---------------------
+
+GOOGLE_CLIENT_ID = "1021957323011-mhbqnueds5ju8mvk1pala345f43arai8.apps.googleusercontent.com"
+client_secrets_file = os.path.join(pathlib.Path(__file__).parent, "client_secret.json")
+
+flow = Flow.from_client_secrets_file(
+    client_secrets_file=client_secrets_file,
+    scopes=["https://www.googleapis.com/auth/userinfo.profile", "https://www.googleapis.com/auth/userinfo.email", "openid"],
+    redirect_uri="http://127.0.0.1:5000/callback"
+)
+
+
+Authorised = ['shivanshsharma8899@gmail.com','abc@gmail.com','pqr@gmail.com']
+
+
+def login_is_required(function):
+    def wrapper(*args, **kwargs):
+        if "google_id" not in session:
+            return abort(401)  # Authorization required
+        else:
+            return function()
+
+    return wrapper
+
+@app.route("/login")
+def login():
+    authorization_url, state = flow.authorization_url()
+    session["state"] = state
+    return redirect(authorization_url)
+
+@app.route("/callback")
+def callback():
+    flow.fetch_token(authorization_response=request.url)
+
+    if not session["state"] == request.args["state"]:
+        abort(500)  # State does not match!
+
+    credentials = flow.credentials
+    request_session = requests.session()
+    cached_session = cachecontrol.CacheControl(request_session)
+    token_request = google.auth.transport.requests.Request(session=cached_session)
+
+    id_info = id_token.verify_oauth2_token(
+        id_token=credentials._id_token,
+        request=token_request,
+        audience=GOOGLE_CLIENT_ID
+    )
+
+    session["email"] = id_info.get("email")
+    session["google_id"] = id_info.get("sub")
+    session["name"] = id_info.get("name")
+    if session["email"] in Authorised:
+        return redirect("/protected_area")
+    return redirect("/logout")
+
+@app.route("/logout")
+def logout():
+    session.clear()
+    return redirect("/")
+
+@app.route("/protected_area")
+@login_is_required
+def protected_area():
+    return f"Hello {session['name']}! <br/> <a href='/logout'><button>Logout</button></a>"
+
+# /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 @app.route("/")
 def home():
